requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-appconf==1.0.6
django-auditlog==3.0.0
django-cors-headers==4.4.0
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-apps-meet==0.1.8
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
multidict==6.1.0
oauthlib==3.2.2
packaging==24.1
pika==1.3.2
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
twilio==9.3.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.11.1
pillow
